0x04-TypeScript/task_1/js/main.ts
// src/main.ts

interface Teacher {
  readonly firstName: string;
  readonly lastName: string;
  fullTimeEmployee: boolean;
  yearsOfExperience?: number;
  location: string;
  [propName: string]: any;
}

const teacher1: Teacher = {
  firstName: 'Alice',
  lastName: 'Johnson',
  fullTimeEmployee: true,
  location: 'New York',
  contract: false,
};

const teacher2: Teacher = {
  firstName: 'Bob',
  lastName: 'Smith',
  fullTimeEmployee: false,
  yearsOfExperience: 10,
  location: 'Los Angeles',
  contract: true,
};

const teachersList: Teacher[] = [teacher1, teacher2];

export { Teacher, teachersList };
